+++ docs/app/layout.tsx
@@ -1,15 +1,12 @@
 import { siteConfig } from '@/lib/config';
 import { b2bTheme, coreTheme, theme } from '@/theme';
 import { MarigoldProvider } from '@/ui';
-import { Suspense } from 'react';
 import '@marigold/theme-b2b/styles.css';
 import '@marigold/theme-core/styles.css';
 import { fontSans } from '@/theme/fonts';
 import { MarigoldThemeSwitch } from '@/ui/ThemeSwitch';
 import { Analytics } from './_components/Analytics';
 import { PortalContaier } from './_components/PortalContainer';
-import { SectionNavigation } from './_components/SectionNavigation';
-import { SiteFooter } from './_components/SiteFooter';
 import { SiteHeader } from './_components/SiteHeader';
 import './globals.css';
 
@@ -21,7 +18,7 @@ const FAV_ICONS = {
 };
 
 export const metadata = {
-  title: 'Marigold Documentation',
+  title: 'Marigold Design System',
   description: "Documentation of Reservix' Design System",
   icons: {
     icon:
@@ -43,41 +40,13 @@ const Layout = ({ children }: { children: React.ReactNode }) => {
   return (
     <html lang="en">
       <body className={`${fontSans.className} min-h-screen`}>
-        <Suspense>
-          <MarigoldThemeSwitch
-            themes={themes}
-            initial={siteConfig.defaultTheme}
-          >
-            <MarigoldProvider theme={theme} className="min-h-screen">
-              <SiteHeader />
-              <aside
-                className={[
-                  'top-[--page-header-height]',
-                  'py-[--page-sub-nav-padding] xl:py-[--page-sub-nav-padding-xl]',
-                  'pl-[--page-padding-md] xl:pl-[--page-padding-xl]',
-                  'h-[calc(100vh-var(--page-header-height))] w-[--page-sub-nav-width] xl:w-[--page-sub-nav-width-xl]',
-                  'fixed z-10 hidden overflow-hidden hover:overflow-y-auto md:block',
-                  'scrollbar-thin scrollbar-thumb-secondary-400 scrollbar-thumb-rounded-full scrollbar-track-transparent',
-                  'border-secondary-200 border-r',
-                ].join(' ')}
-              >
-                {/* current section navigation sidebar */}
-                <SectionNavigation />
-              </aside>
-              <main
-                className={[
-                  'pt-[--page-main-padding] xl:pt-[--page-main-padding-xl]',
-                  'px-[--page-padding] md:px-[--page-padding-md] xl:pr-[--page-padding-xl]',
-                  'md:pl-[calc(var(--page-sub-nav-width)+var(--page-main-padding))] xl:pl-[calc(var(--page-sub-nav-width-xl)+var(--page-main-padding-xl))]',
-                ].join(' ')}
-              >
-                {children}
-                <SiteFooter />
-              </main>
-            </MarigoldProvider>
-            <PortalContaier />
-          </MarigoldThemeSwitch>
-        </Suspense>
+        <MarigoldThemeSwitch themes={themes} initial={siteConfig.defaultTheme}>
+          <MarigoldProvider theme={theme} className="min-h-screen">
+            <SiteHeader />
+            {children}
+          </MarigoldProvider>
+          <PortalContaier />
+        </MarigoldThemeSwitch>
         <Analytics />
       </body>
     </html>
